Limit the maximum number of concurrent diffs
This patch also adds extra diagnostics to improve other errors like this
in the future, and limits the number of blocking threads available to
tokio to reduce the risk of unbounded concurrency in the future.

Fixes #204
